+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [1.1.0]
+
+### Added
+  - Added a configuration option to enable/disable coverage in editor. Default is enabled.
+  - Added a configuration option to enable/disable coverage in gutter. Default is disabled.
+
 ## [1.0.7]
 
 ### Fixed